private network vlan-id not kept after migration
This is for the issue SAIC-89.

CF sometimes uses wrong Segmentation ID, causing public network
creation failure.

The reason for this issue is that CF does not record segmentation ID
for internal network. Instead, it uses next the available one. If a
privat network is created before public network, it will take the ID
1000. If the source cloud's public network has vlan-id 1000, then the
following error will occurs.

To fix this, we record also the old vlan-id used by internal networks
in the source cloud, so there won't be a conflict during migration.

Updated the test_neutron.py to include 'router:external',
'provider:physical_network', and 'provider:network_type'
in network_info.

	modified:   cloudferrylib/os/network/neutron.py

	modified:   cloudferrylib/os/network/neutron.py