SAIC-95: Add flag 'all_vms' to config
Add availability to migrate all VMs with flag 'all_vms=True' in the
'migrate' section of config. It is possible if destination user,
specified in the 'dst' section of config, has admin role in all
destination tenants, otherwise it will fail with 'Unauthorized' error
(Actually, it was required by CloudFerry before to migrate instances
to the different from 'admin' tenant).