Minor fix: Disable 'scp' strict host key checking
Fix small issue in the bandwidth check. Currently, if there is no source
and destination cloud's public ssh keys in the '~/.ssh/known_hosts' of
the CloudFerry node, we have a prompt for the user to confirm the
establishment of connection to both of clouds to perform the bandwidth
check. This fix avoid this promt and adds cloud's public ssh keys to the
'known_hosts' automatically.