Minor fix: Use destination user as default for keypairs
If keypair was created with user, that has been deleted, we use source
admin user (that is specified in the config) as a default for such
keypairs. But in this case we can face up with 2 situations:
 - when source and destination admin users, specified in the config, are
 not the same;
 - If there is no such admin user on the destination, as on source.

In these cases destination admin user, specified in the config, will be
set up as default for such keypairs.

Also, unittests have been provided for this functional.