Minor fix: Disable 'scp' strict host key checking
Fix small issue in the MysqlDump. Currently, if there is no destination
cloud's public ssh key in the '~/.ssh/known_hosts' of the CloudFerry
node, we have a prompt for the user to confirm the establishment of
connection to the cloud to perform MysqlDump action. This fix avoids
this prompt and adds cloud's public ssh key to the 'known_hosts'
automatically.