Added filtering procedure verification.
Updated generate load in order to generate filter file, and create
images with more than one member. Updated config.py file with data
required for this scenario. Added one more VM which will not be
migrated in order to verify filtering procedure by instances.
Updated test cases related to testing of resources that will be or
will not be migrated according to filtering procedure. Added a
file with migration utils methods.