SAIC-312 Floating IP migration
Allows floating IP migration with IP addresses reserved.

Process:
 1. Create floating IP using neutron APIs;
 2. Update neutron database using SQL with correct address.

When admin user gets added to a tenant as member, all it's current
tokens get revoked. This is the primary reason for replacing all the
openstack services client code to not use single token, but generate one
on each operation instead.