[SAIC-371] Fix issue with obsolete images
Remove information about obsolete/broken images (and snapshots) from
`info` variable, because we do not migrate them. We need to delete it,
cause this information is used during the migration process (for example
for images mapping) and can break the migration.

Also, logging has been removed from the `get_image_ref` method of
`glance_image` resource due to it's uselessness.