[SAIC-477] Add floating IPs check
Add floating IPs checking to the `check_networks` action. Patch provides
handling duplicates in floating IPs. Check if floating IP overlaps with
the destination cloud. If floating IP is exactly the same as on the
source cloud (checking by address, tenant, network) check passes. If
difference have been found - exception  will be raised.

Also check if floating IP is not busy (i.e. is not associated to VM on
SRC and DST at the same time) on both environments.