SAIC-737 Fix get_mac_by_ip method for neutron
In neutron two or more ports can have the same IP address, so it's
incorrect to search for MAC address based only on IP address of port.
This patch makes get_mac_by_ip method to only consider ports that are
connected to VM for which we need to find MAC address.