"Behind the proxy" support for devlab setup
As per current design of Cloudferry, Vagrantfile does not considers
'host behind proxy' problem.

If one is trying to setup Cloudferry devlabs, booted VM will have no Internet
access, as Vagrantfile does consider proxy considerations.

Again, with no internet access it not be able to downloads required packages.

When I tried 'vagrant up juno', devstack was not able to download.
And this was basically because of lack of support of 'behind the proxy'.

This patch eliminates this problem.