SAIC-753 Add allowed_address_pairs migration
Modify PrepareNetworks so that it can migrate allowed_address_pairs
property of ports that are connected to VMs.

New task named CreatePortsForVRRP were alse added. The purpose of this
task is to create ports for IP addresses that are listed in
allowed_address_pairs list just to occupy IP address so that some new
VM or DHCP agent will not conflict with virtual IP used in VRRP.