SAIC-788 Fallback to host's IP address if `ext_cidr` doesn't match anything
We often faced an issue when user-configured `ext_cidr` option doesn't
match any IP address on a node, which resulted in debug messages like these:

```
DEBUG remote_runner.py:62 "running 'mktemp -d' on 'None' host as user 'root'"
```

This is happening because `ext_cidr` config option doesn't match any IP
address on a node. In order to avoid this, a fallback path was implemented:
if `ext_cidr` doesn't match any IP address on a node, show warning and
return IP address which was used to reach to a particular node.