[SAIC-646] Remove excess warnings from log
Client proxy catch all exceptions and retry request after waiting for
some time. The problem is that sometimes exception is expected to
occur but proxy client don't know anything about this. So it try to
call API several times before allowing exception to be handled
elsewhere. This results in several WARNINGs shown in log. This
patch fix it by adding special method that can be used to tell
proxy client which exceptions it shouldn't handle.