[CF-102] Add physical networks check
Add 2 types of physnets checks:

1. Check busy physical networks on DST of FLAT network type.
We need this check because FLAT network can be created with one physnet
only once in Neutron. Every next attempt to use the same physnet for
network of such type will be failed.

2. Check physical networks existence on DST for VLAN network type.
We need this check because VLAN network can not be created with physnet,
that does not exist on the terget environment. Every such network
creation attempt will be failed in Neutron.

So we need to prevent these cases and inform about them in the
`CheckNetworks` step at the preparation part of CloudFerry scenario.