[CF-332] Migrate config_drive instance parameter.
When instance is created in cloud with config_drive = True option
in nova.conf, VM instances don't automatically get config_drive
parameter set to True on creation. This means that when migrating
VMs we need to check if drive.config is attached to instance in
order to find out which config_drive value to set on instance
creation.