[CF-270] Create black list feature for images in filter file
Add a new option in the section 'image' in the filter config file
- exclude_images_list. This option allow to specify images list
that will be excluded during the migration process. All other
images will be migrated.
If the 'images_list' and 'exclude_images_list' were specified
exception will be raised. It means that in the filter config file
must exists only one option - 'images_list' or 'exclude_images_list'.