[CF-357] Migrate resources with filter by default
Remove default functionality of migrating whole cloud, if filter is
absent or empty. If the filter file is empty or if the filter file can
not be found or read, CloudFerry will raise an exception, unless there
is an option in the migrate section of CF config file to the effect of
`migrate_whole_cloud = True`.

Make `CheckFilter` return all invalid values (IDs of tenants, volumes,
VMs or images) in the filter config file instead of simple fail on the
first one.