CF-384 Flavor access list for deleted flavors
In case there is more than one flavors with the same ID in source
cloud one of which was deleted previously, migration fails to migrate
flavor which is required for the VM to boot. This patch resolves the
issue by retrieving flavor details from DB in case APIs return
reference to deleted flavor.