[CF-429] Fix cinder migration using non-root user
When doing command piping like this: 'dd of=in_file | ssh user@host
dd of=out_file', it will not work if user don't have access to this
file. Also, we can't use sudo since it's in different cloud and
password can differ from password on source cloud. So the only way
to implement this without terrible hacks is through nc.