CF-472 local_sudo_password pre-check
`local_sudo_password` config option used to introduce a lot of
mistakes into config file because it was unclear whether it's needed
or not and often migration failed with cryptic errors during ISCSI
discovery. This patch adds verification which suggests to set
`local_sudo_password` if it's not set and adds better error handling
for local command execution to avoid cryptic errors.