CF-519 Missing write permissions in destination for file copy
When volumes or ephemerals are copied between compute nodes in source
and destination with non-root user, there might be no write
permissions in the destination folder, which leads to copy failures.
This patch resolves the issue by temporarily adding 777 permissions
for a folder in destination and reverting it back once copy is done.