CF-591 Check if bbcp is installed and allowed to be owerwritten
If migration is done with bbcp as copy engine, bbcp is being copied to
both source and destination nodes prior to ephemeral migration, but if
bbcp was previously copied to the node from different user, current
user may not have permissions to overwrite existing bbcp executable and
migration would fail. This patch resolves this issue by reusing existing
bbcp and assigning 777 permissions to the file so that it can be
overwritten by anyone.